// domains/core-platform-services/auth-service/src/modules/auth/roles/roles.enum.ts
export enum RolesEnum {
  USER = 'USER',
  ADMIN = 'ADMIN',
  ACADEMY_ADMIN = 'ACADEMY_ADMIN',
  ACADEMY_INSTRUCTOR = 'ACADEMY_INSTRUCTOR',
  ACADEMY_STUDENT = 'ACADEMY_STUDENT',
  CONSULTANCY_ADVISOR = 'CONSULTANCY_ADVISOR',
  CONSULTANCY_MANAGER = 'CONSULTANCY_MANAGER',
  CONSULTANCY_CLIENT = 'CONSULTANCY_CLIENT',
  CONTECH_DEVELOPER = 'CONTECH_DEVELOPER',
  CONTECH_DESIGNER = 'CONTECH_DESIGNER',
  CONTECH_PROJECT_MANAGER = 'CONTECH_PROJECT_MANAGER',
  EVENTS_ORGANIZER = 'EVENTS_ORGANIZER',
  EVENTS_PARTICIPANT = 'EVENTS_PARTICIPANT',
  EVENTS_SPONSOR = 'EVENTS_SPONSOR',
}